requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
pandas
matplotlib
python-dotenv